+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Added analog filter option
 - Added ADS111X source dialog
 - Added manual range tuning option for local analog and ADS111X
+- Automatically connects at startup if one supported FFBoard device is found
 
 #### Updater:
 - Added update browser (Help->updates)